backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
google-genai==0.3.0
pydantic==2.8.2
transformers==4.36.0
torch==2.1.0
google-api-python-client==2.108.0

# Web Scraping Dependencies
google-search-results==2.4.2
trafilatura==1.6.2
beautifulsoup4==4.12.2
requests==2.31.0

# Embeddings and Vector DB
tiktoken==0.5.1
fastembed==0.2.2
faiss-cpu==1.7.4
numpy==1.26.2

# ML Model Dependencies (Depression Prediction)
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
